fix: el2 to el1 not working on board
