+++ bins/ppos/src/exception/exception_handler.rs
@@ -2,16 +2,10 @@
 //
 // Copyright (c) 2018-2023 Andre Richter <andre.o.richter@gmail.com>
 
-use aarch64_cpu::{asm::barrier, registers::*};
-use core::{arch::global_asm, cell::UnsafeCell, fmt};
+use aarch64_cpu::registers::*;
+use core::fmt;
 use library::println;
-use tock_registers::{
-    interfaces::{Readable, Writeable},
-    registers::InMemoryRegister,
-};
-
-// Assembly counterpart to this file.
-global_asm!(include_str!("exception.s"));
+use tock_registers::{interfaces::Readable, registers::InMemoryRegister};
 
 //--------------------------------------------------------------------------------------------------
 // Private Definitions
@@ -261,46 +255,4 @@ impl fmt::Display for ExceptionContext {
         }
         write!(f, "      lr : {:#018x}", self.lr)
     }
-}
-
-//--------------------------------------------------------------------------------------------------
-// Public Code
-//--------------------------------------------------------------------------------------------------
-#[derive(Eq, PartialEq)]
-pub enum PrivilegeLevel {
-    User,
-    Kernel,
-    Hypervisor,
-    Unknown,
-}
-
-/// The processing element's current privilege level.
-pub fn current_privilege_level() -> (PrivilegeLevel, &'static str) {
-    let el = CurrentEL.read_as_enum(CurrentEL::EL);
-    match el {
-        Some(CurrentEL::EL::Value::EL2) => (PrivilegeLevel::Hypervisor, "EL2"),
-        Some(CurrentEL::EL::Value::EL1) => (PrivilegeLevel::Kernel, "EL1"),
-        Some(CurrentEL::EL::Value::EL0) => (PrivilegeLevel::User, "EL0"),
-        _ => (PrivilegeLevel::Unknown, "Unknown"),
-    }
-}
-
-/// Init exception handling by setting the exception vector base address register.
-///
-/// # Safety
-///
-/// - Changes the HW state of the executing core.
-/// - The vector table and the symbol `__exception_vector_table_start` from the linker script must
-///   adhere to the alignment and size constraints demanded by the ARMv8-A Architecture Reference
-///   Manual.
-pub unsafe fn handling_init() {
-    // Provided by exception.S.
-    extern "Rust" {
-        static __exception_vector_start: UnsafeCell<()>;
-    }
-
-    VBAR_EL1.set(__exception_vector_start.get() as u64);
-
-    // Force VBAR update to complete before next instruction.
-    barrier::isb(barrier::SY);
 }
