+++ libs/cpu/Cargo.toml
@@ -8,4 +8,3 @@ edition = "2021"
 [dependencies]
 aarch64-cpu = "9.4.0"
 tock-registers = "0.8.1"
-library = { path = "../library" }
