+++ libs/cpu/src/lib.rs
@@ -1,4 +1,3 @@
 #![no_std]
 
 pub mod cpu;
-pub mod exception;
